improve app structure, fix swiftlint warnings
