+++ App/CourseApp/App/Navigation/NavigationControllerCoordinator.swift
@@ -1,0 +1,18 @@
+//
+//  NavigationControllerCoordinator.swift
+//  CourseApp
+//
+//  Created by Marcel Mravec on 26.05.2024.
+//
+
+import UIKit
+
+protocol NavigationControllerCoordinator: ViewControllerCoordinator {
+    var navigationController: UINavigationController { get }
+}
+
+extension NavigationControllerCoordinator {
+    var rootViewController: UIViewController {
+        navigationController
+    }
+}
