+++ App/CourseApp/App/Navigation/TabBarControllerCoordinator.swift
@@ -1,0 +1,18 @@
+//
+//  TabBarControllerCoordinator.swift
+//  CourseApp
+//
+//  Created by Marcel Mravec on 26.05.2024.
+//
+
+import UIKit
+
+protocol TabBarControllerCoordinator: ViewControllerCoordinator {
+    var tabBarController: UITabBarController { get }
+}
+
+extension TabBarControllerCoordinator {
+    var rootViewController: UIViewController {
+        tabBarController
+    }
+}
