+++ App/CourseApp/App/Navigation/ViewControllerCoordinator.swift
@@ -1,0 +1,12 @@
+//
+//  ViewControllerCoordinator.swift
+//  CourseApp
+//
+//  Created by Marcel Mravec on 24.05.2024.
+//
+
+import UIKit
+
+protocol ViewControllerCoordinator: Coordinator {
+    var rootViewController: UIViewController { get }
+}
