+++ App/CourseApp/App/Resources/Colors.xcassets/Contents.json
