+++ App/CourseApp/App/Resources/Colors.xcassets/bgColor.colorset/Contents.json
