+++ App/CourseApp/App/Resources/Images.xcassets/AccentColor.colorset/Contents.json
