+++ App/CourseApp/App/Resources/Images.xcassets/AppIcon.appiconset/Contents.json
