+++ App/CourseApp/App/Resources/Images.xcassets/Contents.json
