+++ App/CourseApp/App/Resources/Images.xcassets/back.imageset/Contents.json
