+++ App/CourseApp/App/Resources/Images.xcassets/computer.imageset/Contents.json
