+++ App/CourseApp/App/Resources/Images.xcassets/computer.imageset/computer.jpg
