--- App/CourseApp/Navigation/ViewControllerCoordinator.swift
@@ -1,32 +1,0 @@
-//
-//  ViewControllerCoordinator.swift
-//  CourseApp
-//
-//  Created by Marcel Mravec on 24.05.2024.
-//
-
-import UIKit
-
-protocol ViewControllerCoordinator: Coordinator {
-    var rootViewController: UIViewController { get }
-}
-
-protocol NavigationControllerCoordinator: ViewControllerCoordinator {
-    var navigationController: UINavigationController { get }
-}
-
-extension NavigationControllerCoordinator {
-    var rootViewController: UIViewController {
-        navigationController
-    }
-}
-
-protocol TabBarControllerCoordinator: ViewControllerCoordinator {
-    var tabBarController: UITabBarController { get }
-}
-
-extension TabBarControllerCoordinator {
-    var rootViewController: UIViewController {
-        tabBarController
-    }
-}
