--- App/CourseApp/Views/UIKit/MainTabBarController.swift
@@ -1,41 +1,0 @@
-//
-//  MainTabBarController.swift
-//  CourseApp
-//
-//  Created by Marcel Mravec on 17.05.2024.
-//
-
-import SwiftUI
-import UIKit
-
-//struct MainTabView: UIViewControllerRepresentable {
-//    func makeUIViewController(context: Context) -> MainTabBarController {
-//        MainTabBarController()
-//    }
-//    
-//    func updateUIViewController(_ uiViewController: MainTabBarController, context: Context) {}
-//}
-
-//class MainTabBarController: UITabBarController {
-//    override func viewDidLoad() {
-//        super.viewDidLoad()
-//        setupGlobalTabBarUI()
-//    }
-//    
-//    func setupGlobalTabBarUI() {
-//        UITabBar.appearance().backgroundColor = .systemBrown
-//        UITabBar.appearance().tintColor = .white
-//    }
-//    
-//    func setupAppearance() {
-//               
-//        UITabBar.appearance().backgroundColor = .systemBrown
-//        UITabBar.appearance().tintColor = .white
-//        UITabBarItem.appearance().setTitleTextAttributes(
-//            [
-//                NSAttributedString.Key.font : TextType.caption.font
-//            ],
-//            for: .normal)
-//        UINavigationBar.appearance().tintColor = .white
-//    }
-//}
